Revert "Измения дб в соответствии с новым тз"

+++ src/main/java/com/service/inspection/repositories/InspectionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InspectionRepository extends JpaRepository<Inspection, Long> {
+public interface InspectionRepository extends JpaRepository<Inspection, Integer> {
 
 }
